WIP: Changed model to handle the int64 version of the model weights

+++ src/main/scala/com/johnsnowlabs/ml/ai/util/Generation/Logit/LogitProcess/MinLengthLogitProcessor.scala
@@ -16,10 +16,10 @@
 
 package com.johnsnowlabs.ml.ai.util.Generation.Logit.LogitProcess
 
-class MinLengthLogitProcessor(val eosTokenId: Int, val minLength: Int, val vocabSize: Int)
+class MinLengthLogitProcessor(val eosTokenId: Long, val minLength: Int, val vocabSize: Int)
     extends LogitProcessor {
   override def call(
-      inputIds: Seq[Array[Int]],
+      inputIds: Seq[Array[Long]],
       scores: Array[Array[Float]],
       currentLength: Int): Array[Array[Float]] = {
     if (!eosTokenId.isNaN && currentLength < this.minLength) {
